feat: App / Get platform info from ReactNative #570
