+++ client/src/layout/Root.tsx
@@ -9,25 +9,10 @@ import { uiActions } from '../store/ui';
 import classes from './Root.module.css';
 
 function Root() {
-  const dispatch = useAppDispatch();
   const location = useLocation();
 
   const { isGuest } = useAppSelector((state) => state.user.auth);
 
-  useEffect(() => {
-    const setPlatform = (event: MessageEvent) => {
-      const { platform } = JSON.parse(event.data);
-
-      if (platform === 'android' || platform === 'iOS') {
-        dispatch(uiActions.setPlatform(platform));
-      }
-    };
-
-    // android
-    // document.addEventListener('message', setPlatform);
-    // ios
-    window.addEventListener('message', setPlatform);
-  });
 
   useEffect(() => {
     if (location.pathname !== '/user') {
